+++ TaskMenu/Views/TaskPopoverViewController.swift
@@ -299,7 +299,7 @@ final class TaskPopoverViewController: NSViewController {
 
     private func openSettings() {
         NSApp.activate(ignoringOtherApps: true)
-        NSApp.sendAction(Selector(("showSettingsWindow:")), to: nil, from: nil)
+        NSApp.sendAction(#selector(TaskMenuAppDelegate.showSettingsWindow(_:)), to: NSApp.delegate, from: nil)
         onRequestClose()
     }